+++ HDF5/H5Ppublic.cs
@@ -966,6 +966,45 @@ namespace HDF.PInvoke
             (hid_t fcpl, ref H5F.file_space_type_t strategy,
             ref hsize_t threshold);
 
+        /// <summary>
+        /// Retrieves the file space page size for a file creation property
+        /// list.
+        /// See https://support.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetFileSpacePageSize
+        /// </summary>
+        /// <param name="fcpl">The file creation property list identifier.</param>
+        /// <param name="strategy">The current file space management strategy
+        /// in use for the file.</param>
+        /// <param name="fsp_size">File space page size</param>
+        /// <returns>Returns a non-negative value if successful; otherwise
+        /// returns a negative value.</returns>
+        [DllImport(Constants.DLLFileName,
+            EntryPoint = "H5Pget_file_space_page_size",
+            CallingConvention = CallingConvention.Cdecl),
+        SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
+        public static extern herr_t get_file_space_page_size
+            (hid_t fcpl, ref hsize_t fsp_size);
+
+        /// <summary>
+        /// Retrieves the file space handling strategy, persisting free-space
+        /// condition and threshold value for a file creation property list.
+        /// See https://support.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetFileSpaceStrategy
+        /// </summary>
+        /// <param name="fcpl">The file creation property list identifier.</param>
+        /// <param name="strategy">The current file space management strategy
+        /// in use for the file.</param>
+        /// <param name="persist">The boolean value indicating whether free
+        /// space is persistent or not.</param>
+        /// <param name="threshold">The free-space section size threshold value.</param>
+        /// <returns>Returns a non-negative value if successful; otherwise
+        /// returns a negative value.</returns>
+        [DllImport(Constants.DLLFileName,
+            EntryPoint = "H5Pget_file_space_strategy",
+            CallingConvention = CallingConvention.Cdecl),
+        SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
+        public static extern herr_t get_file_space_strategy(hid_t fcpl,
+            ref H5F.fspace_strategy_t strategy, ref hbool_t persist,
+            ref hsize_t threshold);
+
 #endif
 
         /// <summary>
@@ -1241,7 +1280,7 @@ namespace HDF.PInvoke
             CallingConvention = CallingConvention.Cdecl),
         SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
         public static extern herr_t get_mdc_config
-            (hid_t plist_id, ref H5AC.cache_config_t config_ptr);
+            (hid_t plist_id, IntPtr config_ptr);
   
 #if HDF5_VER1_10
 
@@ -1260,7 +1299,7 @@ namespace HDF.PInvoke
            CallingConvention = CallingConvention.Cdecl),
         SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
         public static extern herr_t get_mdc_image_config
-            (hid_t fapl_id, ref H5AC.cache_image_config_t config_ptr);
+            (hid_t fapl_id, IntPtr config_ptr);
 
         /// <summary>
         /// Gets metadata cache logging options.
@@ -1393,6 +1432,28 @@ namespace HDF.PInvoke
         SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
         public static extern herr_t get_object_flush_cb
             (hid_t fapl_id, ref H5F.flush_cb_t func, ref IntPtr udata);
+
+        /// <summary>
+        /// Retrieves the maximum size for the page buffer and the minimum
+        /// percentage for metadata and raw data pages.
+        /// See https://support.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetPageBufferSize
+        /// </summary>
+        /// <param name="fapl_id">File access property list identifier</param>
+        /// <param name="buf_size">Maximum size, in bytes, of the page buffer</param>
+        /// <param name="min_meta_prec">Minimum metadata percentage to keep in
+        /// the page buffer before allowing pages containing metadata to be
+        /// evicted</param>
+        /// <param name="min_raw_perc">Minimum raw data percentage to keep in
+        /// the page buffer before allowing pages containing raw data to be
+        /// evicted</param>
+        /// <returns>Returns a non-negative value if successful; otherwise
+        /// returns a negative value.</returns>
+        [DllImport(Constants.DLLFileName, EntryPoint = "H5Pget_page_buffer_size",
+            CallingConvention = CallingConvention.Cdecl),
+        SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
+        public static extern herr_t get_page_buffer_size
+            (hid_t fapl_id, ref IntPtr buf_size, ref uint min_meta_perc,
+            ref uint min_raw_perc);
 
 #endif
 
@@ -2625,6 +2686,44 @@ namespace HDF.PInvoke
         public static extern herr_t set_file_space
             (hid_t fcpl, H5F.file_space_type_t strategy, hsize_t threshold = 1);
 
+        /// <summary>
+        /// Sets the file space page size for a file creation property list.
+        /// See https://support.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-SetFileSpacePageSize
+        /// </summary>
+        /// <param name="fcpl">The file creation property list identifier.</param>
+        /// <param name="strategy">The current file space management strategy
+        /// in use for the file.</param>
+        /// <param name="fsp_size">File space page size</param>
+        /// <returns>Returns a non-negative value if successful; otherwise
+        /// returns a negative value.</returns>
+        [DllImport(Constants.DLLFileName,
+            EntryPoint = "H5Pset_file_space_page_size",
+            CallingConvention = CallingConvention.Cdecl),
+        SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
+        public static extern herr_t set_file_space_page_size
+            (hid_t fcpl, hsize_t fsp_size);
+        
+        /// <summary>
+        /// Sets the file space handling strategy and persisting free-space
+        /// values for a file creation property list.
+        /// See https://support.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-SetFileSpaceStrategy
+        /// </summary>
+        /// <param name="fcpl">The file creation property list identifier.</param>
+        /// <param name="strategy">The file space handling strategy to be used.</param>
+        /// <param name="persist">A boolean value to indicate whether free
+        /// space should be persistent or not.</param>
+        /// <param name="threshold">The smallest free-space section size that
+        /// the free space manager will track.</param>
+        /// <returns>Returns a non-negative value if successful; otherwise
+        /// returns a negative value.</returns>
+        [DllImport(Constants.DLLFileName,
+            EntryPoint = "H5Pset_file_space_strategy",
+            CallingConvention = CallingConvention.Cdecl),
+        SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
+        public static extern herr_t set_file_space_strategy(hid_t fcpl,
+            H5F.fspace_strategy_t strategy, hbool_t persist,
+            hsize_t threshold);
+
 #endif
 
         /// <summary>
@@ -2871,7 +2970,7 @@ namespace HDF.PInvoke
             CallingConvention = CallingConvention.Cdecl),
         SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
         public static extern herr_t set_mdc_config
-            (hid_t plist_id, ref H5AC.cache_config_t config_ptr);
+            (hid_t plist_id, IntPtr config_ptr);
 
 #if HDF5_VER1_10
 
@@ -2889,7 +2988,7 @@ namespace HDF.PInvoke
             CallingConvention = CallingConvention.Cdecl),
         SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
         public static extern herr_t set_mdc_image_config
-            (hid_t fapl_id, ref H5AC.cache_image_config_t config_ptr);
+            (hid_t fapl_id, IntPtr config_ptr);
 
         /// <summary>
         /// Sets metadata cache logging options.
@@ -3008,6 +3107,28 @@ namespace HDF.PInvoke
         SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
         public static extern herr_t set_object_flush_cb
             (hid_t plist_id, H5F.flush_cb_t func, IntPtr udata);
+
+        /// <summary>
+        /// Sets the maximum size for the page buffer and the minimum
+        /// percentage for metadata and raw data pages.
+        /// See https://support.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-SetPageBufferSize
+        /// </summary>
+        /// <param name="fapl_id">File access property list identifier</param>
+        /// <param name="buf_size">Maximum size, in bytes, of the page buffer</param>
+        /// <param name="min_meta_prec">Minimum metadata percentage to keep in
+        /// the page buffer before allowing pages containing metadata to be
+        /// evicted</param>
+        /// <param name="min_raw_perc">Minimum raw data percentage to keep in
+        /// the page buffer before allowing pages containing raw data to be
+        /// evicted</param>
+        /// <returns>Returns a non-negative value if successful; otherwise
+        /// returns a negative value.</returns>
+        [DllImport(Constants.DLLFileName, EntryPoint = "H5Pset_page_buffer_size",
+            CallingConvention = CallingConvention.Cdecl),
+        SuppressUnmanagedCodeSecurity, SecuritySafeCritical]
+        public static extern herr_t set_page_buffer_size
+            (hid_t fapl_id, IntPtr buf_size, uint min_meta_perc,
+            uint min_raw_perc);
 
 #endif