+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+##### 1.10.1.0
+* Updated native dependencies (HDF5 1.10.1) 
+* New functions for metadata cache image handling
+* MDC H5P.set_evict_on_close
+* New functions for file space management (page buffering, paged aggr.)
+
 ##### 1.8.19.0
 * Updated native dependencies (HDF5 1.8.19) 
 * New H5PL functions